........
  r33626 | mindless | 2010-08-07 22:29:34 -0700 (Sat, 07 Aug 2010) | 13 lines
  
  [SECURITY-5] Add several missing permission checks in CLI commands.
  Current CLI is always available and most commands can be run by anonymous users.
  Now only a short help message and the login command are available for anonymous
  in a Hudson without anonymous read permission, and each individual command also
  checks the appropriate permission for that command.  Details:
  * Vulnerable CLI commands now fixed:
    build, clear-queue, copy-job, create-job, delete-builds, delete-job,
    disable-job, enable-job, install-plugin, install-tool, keep-build,
    mail, restart, safe-restart, set-build-result
  * Broken CLI commands now fixed:
    connect-node, delete-node, disconnect-node, offline-node, online-node
  * Other vulnerable command (not CLI) now fixed: /gc (JVM garbage collect)
........


git-svn-id: https://hudson.dev.java.net/svn/hudson/branches/rc@33641 71c3de6d-444a-0410-be80-ed276b4c234a

+++ core/src/main/java/hudson/cli/DeleteBuildsCommand.java
@@ -1,7 +1,7 @@
 /*
  * The MIT License
  *
- * Copyright (c) 2004-2009, Sun Microsystems, Inc.
+ * Copyright (c) 2004-2010, Sun Microsystems, Inc.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -25,6 +25,7 @@ package hudson.cli;
 
 import hudson.Extension;
 import hudson.model.AbstractBuild;
+import hudson.model.Run;
 
 import java.io.IOException;
 import java.io.PrintStream;
@@ -51,6 +52,8 @@ public class DeleteBuildsCommand extends AbstractBuildRangeCommand {
 
     @Override
     protected int act(List<AbstractBuild<?, ?>> builds) throws IOException {
+        job.checkPermission(Run.DELETE);
+
         for (AbstractBuild build : builds)
             build.delete();
